`decoration`: Fix incorrect ideas, set functions as-needed.
Prior to this patch, decorations and functions were muddled together and
sent together. I'm not sure why I thought that was at all correct.

This implementation correctly allows for all forms of decoration.
However, it leaves functions alone. This means that decorations can be
restricted, but functions cannot be. Which, in turn, means that
`decoration none` will result in an undecorated window that can be moved
and resized (assuming wm support for alt+mouse clicking).

I'm open to allowing this command to explicitly specify functions as
well, but I'll save that for when a sufficient use case is provided. I
like being able to move and resize my borderless windows.